Crate markdown_composer

Re-exports

pub use crate::builders::image::ImageBuilder;

pub use crate::builders::link::LinkBuilder;

pub use crate::builders::list::ListBuilder;

pub use crate::traits::MarkdownElement;

pub use crate::transforms::Bold;

pub use crate::types::header::Header;

pub use crate::types::header::HeaderLevel;

pub use crate::types::image::Image;

pub use crate::types::link::Link;

pub use crate::types::list::List;

pub use crate::types::list::ListItem;

pub use crate::types::markdown::Markdown;

pub use crate::types::paragraph::Paragraph;

pub use crate::extensions::github::CheckmarkItem;

Modules

builders
Contains builders for all Markdown content types.

extensions
This provide provides certain markdown feature extensions to the crate that can be enabled through crate features.

traits
Contains Markdown trait definitions.

transforms
Contains extension traits that can used to transform text elements into special markdown formats.

types
This module defines all Markdown types that are specified inside the CommonMark specification.