Trait markdown_composer::traits::AsFooter

pub trait AsFooter {
    fn as_footer(&self) -> Box<dyn MarkdownElement>;
}

A MarkdownElement that can be rendered as a footer value.

Required methods

fn as_footer(&self) -> Box<dyn MarkdownElement>

Returns the MarkdownElement that can be used to render the footer.

Loading content...